% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medFiltImages.R
\docType{data}
\name{medFiltImages}
\alias{medFiltImages}
\title{Neighbor-median filtered test images}
\format{
An object of class \code{'list'};
}
\usage{
data(medFiltImages)
}
\description{
The two images in the testImages file that has gone through the neighMedFilt
function. Data from Ramberger et al, Brain, 2020.
}
\keyword{datasets}
