% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantCutoff.R
\name{quantCutoff}
\alias{quantCutoff}
\title{Find a quantile-based cutoff to use for images or movies}
\usage{
quantCutoff(ctrlImageList, quant = 0.99)
}
\arguments{
\item{ctrlImageList}{Either a named list, as with the
example file or a movie, or a vector of RGB png file names, e.g. generated with
list.files(). The second option is preferred for images, as this is a more
reasonable method from a memory perspective.}

\item{quant}{The quantile of pixel intensities in all the images in
ctrlImageList from which the cutoff value will be derived.}
}
\value{
The median value for the quantile specified in quant for all the
images in the ctrlImageList.
}
\description{
Here, a set of control frames and a quantile is provided, which together
generate a cutoff used for downstream analyses.
}
\examples{

#Load example data.
data(testImages)

#Run the function with the output in the current directory
quantCutoff(testImages[1], 0.99)
#Expected result: 0.1098039
}
