% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testImages.R
\docType{data}
\name{testImages}
\alias{testImages}
\title{Test images}
\format{
An object of class \code{'list'};
}
\usage{
data(testImages)
}
\description{
Two images in 2D matrix format rendered from monochromatic pngs with the png
package. Data from Ramberger et al, Brain, 2020.
}
\keyword{datasets}
