% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testMovie.R
\docType{data}
\name{testMovie}
\alias{testMovie}
\title{Test movie}
\format{
An object of class \code{'list'};
}
\usage{
data(testMovie)
}
\description{
A 12-frame movie sampled from Ramberger et al, Brain, 2020. The four first
frames come from the first hour, and after that, one frame comes every 6.25
hours.
}
\keyword{datasets}
