% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medMadNormSum.R
\name{medMadNormSum}
\alias{medMadNormSum}
\title{Normalize each movie frame to its median and MAD and sum the result for all
pixels}
\usage{
medMadNormSum(imageList, quantThreshFrames = 1:4, outDir = "default")
}
\arguments{
\item{imageList}{A named list of images, as with the
example file or a movie, or a vector of TIFF file names, e.g. generated with
list.files().¨}

\item{quantThreshFrames}{The frames in the movie that will be used to set the
quantile-based threshold for summation. Default is frame 1:4, as four
pictures were taken per hour in the paper.}

\item{outDir}{If the imageList is a vector of TIFF file names, this will
be the place where the filtered movies will be saved. Default is to save
a new .rds file in the same directory as the old file, but with the extension
"normalized". Can}
}
\value{
A list of two, both containing lists of length(imageList), containing
1: pixel intensity sums for each frame in above the threshold.
2: quantiles that should be used as input for the normMovie function.
In addition, side effect in form of filtered movies, one per imageList item.
890:950,1285:1325
}
\description{
This function is meant to normalize frames of a movie, as there might be
slight variations in light intensity etc over the course of a live-cell
experiment lasting days.
}
\examples{

#Load example data.
data(testImages)

#Run the function with the output in the current directory
medMadNormIm <- medMadNorm(testImages)

}
