% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixSum.R
\name{pixSum}
\alias{pixSum}
\title{Sum pixel intensity above threshold}
\usage{
pixSum(imageList, quantCutThreshold)
}
\arguments{
\item{imageList}{Either a named list of images, as with the
example file or a movie, or a vector of RGB png file names, e.g. generated with
list.files(). The second option is preferred for images, as this is a more
reasonable method from a memory perspective.}

\item{quantCutThreshold}{A result from the quantCutoff function on a relevant
control dataset. NB! The same controls can be included in the imageList too.}
}
\value{
A named vector with the sum of pixel intensities above the
quantCutThreshold normalized to the total number of pixels in the frame.
}
\description{
In this function, we use the input from the quantCutoff function to sum the
pixel intensity in the fraction above the threshold. This is the
central function of the package.
}
\examples{

#Load example data.
data(testImages)

#Run the function with the output in the current directory
pixSum(testImages, quantCutoff(testImages[1], 0.99))
#Expected result:
#       A33        P79
#0.00140716 0.02007561
}
